Models (2)/oral_cancer_detection_app/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
numpy==1.24.3
Pillow==10.0.0
gunicorn==21.2.0
scikit-learn==1.3.0
tensorflow==2.13.0
opencv-python==4.8.0.76 